We have slicing, and it is fast!
There is a lot of nasty to fix, but we're getting there.
The STL lib has bugs in both ascii and binary mode, but I
was able to fix the broken ascii decoder. Binary is a mystery
that I'll fix later on. We CAN however, slice some fairly
complicated models. The AABB tree is building good results
and is able to slice the 3d model of my head without errors,
which is pretty amazing since it is from pointcloud data.

Big TODO:
 1. Fix the debug output, there is too much of it in the tests and libs
 2. Fix binary stl decoding.
 3. Build a command line client that can actually take the slices and
    generate toolpaths and/or svg and/or png output.
 4. Infill infill infill.