W007, slicing now works in the gui.
Turns out that my polygonizer returned a generator (wups) which
is awesome because lazy eval, but bad because it only worked on
the first cycle through, which meant that it could only be
consumed once; and debugging it resulted in a heisenbug.

Now I need to add perimeter clipping and insetting on the initial
perims, but this is a good first start.